requirements.txt
amqp==5.2.0
asgiref==3.6.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.13.1
flower==2.0.1
fsspec==2024.2.0
gunicorn==20.1.0
humanize==4.9.0
idna==3.6
inflection==0.5.1
Jinja2==3.1.3
kombu==5.3.5
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
mysqlclient==2.1.1
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.6
pydub==0.25.1
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.6.0
torch==2.2.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
